Group 'command' API, rename 'general' to 'server' api
