--- yamcs-api/src/main/proto/yamcs/protobuf/commanding/command_history_service.proto
@@ -1,98 +1,0 @@
-syntax="proto2";
-
-package yamcs.protobuf.commanding;
-
-option java_package = "org.yamcs.protobuf";
-option java_outer_classname = "CommandHistoryServiceProto";
-option java_multiple_files = true;
-
-import "google/protobuf/timestamp.proto";
-
-import "yamcs/api/annotations.proto";
-import "yamcs/protobuf/commanding/commanding.proto";
-
-service CommandHistoryApi {
-
-  // List commands
-  rpc ListCommands(ListCommandsRequest) returns (ListCommandsResponse) {
-    option (yamcs.api.route) = {
-      get: "/api/archive/{instance}/commands"
-    };
-  }
-  
-  // Get a command
-  rpc GetCommand(GetCommandRequest) returns (CommandHistoryEntry) {
-    option (yamcs.api.route) = {
-      get: "/api/archive/{instance}/commands/{id}"
-    };
-  }
-  
-  // Streams back commands
-  rpc StreamCommands(StreamCommandsRequest) returns (stream CommandHistoryEntry) {
-    option (yamcs.api.route) = {
-      post: "/api/stream-archive/{instance}:streamCommands"
-      body: "*"
-    };
-  }
-  
-  // Receive updates on issued commands
-  rpc SubscribeCommands(SubscribeCommandsRequest) returns (stream CommandHistoryEntry) {
-    option (yamcs.api.websocket) = {
-      topic: "commands"
-    };
-  }
-}
-
-message ListCommandsRequest {
-  // Yamcs instance name.
-  optional string instance = 1;
-  
-  // The zero-based row number at which to start outputting results.
-  // Default: ``0``
-  optional int64 pos = 2;
-  
-  // The maximum number of returned records per page. Choose this value too
-  // high and you risk hitting the maximum response size limit enforced by
-  // the server. Default: ``100``
-  optional int32 limit = 3;
-  
-  // The order of the returned results. Can be either ``asc`` or ``desc``.
-  // Default: ``desc``
-  optional string order = 4;
-  
-  // Text to search in the name of the command.
-  optional string q = 5;
-
-  optional string next = 6;
-  
-  // Filter the lower bound of the command's generation time. Specify a date
-  // string in ISO 8601 format. This bound is inclusive.
-  optional google.protobuf.Timestamp start = 7;
-  
-  // Filter the upper bound of the command's generation time. Specify a date
-  // string in ISO 8601 format. This bound is exclusive.
-  optional google.protobuf.Timestamp stop = 8;
-}
-
-message ListCommandsResponse {
-  repeated CommandHistoryEntry entry = 1;
-  optional string continuationToken = 2;
-}
-
-message GetCommandRequest {
-  optional string instance = 1;
-  optional string id = 2;
-}
-
-message StreamCommandsRequest {
-  optional string instance = 1;
-  optional google.protobuf.Timestamp start = 2;
-  optional google.protobuf.Timestamp stop = 3;
-  repeated string name = 4;
-}
-
-message SubscribeCommandsRequest {
-  optional string instance = 1;
-  optional string processor = 2;
-  optional bool ignorePastCommands = 3;
-}
