+++ yamcs-core/src/main/java/org/yamcs/http/HttpServer.java
@@ -36,10 +36,9 @@ import org.yamcs.http.api.BucketsApi;
 import org.yamcs.http.api.CfdpApi;
 import org.yamcs.http.api.ClearanceApi;
 import org.yamcs.http.api.ClientsApi;
-import org.yamcs.http.api.CommandHistoryApi;
+import org.yamcs.http.api.CommandsApi;
 import org.yamcs.http.api.Cop1Api;
 import org.yamcs.http.api.EventsApi;
-import org.yamcs.http.api.GeneralApi;
 import org.yamcs.http.api.IamApi;
 import org.yamcs.http.api.IndexesApi;
 import org.yamcs.http.api.ManagementApi;
@@ -49,6 +48,7 @@ import org.yamcs.http.api.ParameterArchiveApi;
 import org.yamcs.http.api.ProcessingApi;
 import org.yamcs.http.api.QueueApi;
 import org.yamcs.http.api.RocksDbApi;
+import org.yamcs.http.api.ServerApi;
 import org.yamcs.http.api.StreamArchiveApi;
 import org.yamcs.http.api.TableApi;
 import org.yamcs.http.api.TagApi;
@@ -243,9 +243,8 @@ public class HttpServer extends AbstractYamcsService {
         addApi(new CfdpApi());
         addApi(new ClearanceApi());
         addApi(new ClientsApi());
-        addApi(new CommandHistoryApi());
+        addApi(new CommandsApi());
         addApi(new Cop1Api());
-        addApi(new GeneralApi(this));
         addApi(new EventsApi());
         addApi(new IamApi());
         addApi(new IndexesApi());
@@ -255,6 +254,7 @@ public class HttpServer extends AbstractYamcsService {
         addApi(new ParameterArchiveApi());
         addApi(new ProcessingApi());
         addApi(new QueueApi());
+        addApi(new ServerApi(this));
         addApi(new StreamArchiveApi());
         addApi(new RocksDbApi());
         addApi(new TableApi());
